backend/requirements.txt
# Backend dependencies for FastAPI
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
motor==3.3.2
redis==5.0.1
python-dotenv==1.0.0
python-multipart==0.0.6
aiofiles==23.2.1
slowapi==0.1.9
# Add other dependencies as needed